+++ R/tagSum.R
@@ -1,11 +1,14 @@
 #' General summary of detections for each tag
 #'
-#' Creates a summary for each tag of it's first and last detection time, first and last detection site,
-#' length of time between first and last detection,  straight line distance between first and last detection site,
-#' rate of movement, and bearing
+#' Creates a summary for each tag of it's first and last detection time, first
+#' and last detection site, length of time between first and last detection,
+#' straight line distance between first and last detection site, rate of
+#' movement, and bearing
 #'
-#' @param data a selected table from .motus data, eg. "alltags", or a data.frame of detection data 
-#' including at a minimum variables for motusTagID, fullID, recvDeployLat, recvDeployLon, recvDeployName, ts
+#' @param data a selected table from .motus data, eg. "alltagsGPS", or a
+#'   data.frame of detection data including at a minimum variables for
+#'   motusTagID, fullID, recvDeployLat, recvDeployLon, recvDeployName, ts,
+#'   gpsLat, gpsLon
 #' @export
 #' @author Zoe Crysler \email{zcrysler@@gmail.com}
 #'
@@ -27,10 +30,11 @@
 #' }
 #'
 #' @examples
-#' # You can use either a selected tbl from .motus eg. "alltags", or a
+#' # You can use either a selected tbl from .motus eg. "alltagsGPS", or a
 #' # data.frame, instructions to convert a .motus file to all formats are below.
 #' 
 #' # download and access data from project 176 in sql format
+#' # usename and password are both "motus.sample"
 #' \dontrun{sql.motus <- tagme(176, new = TRUE, update = TRUE)}
 #' 
 #' # use example sql file included in `motus`
@@ -39,7 +43,7 @@
 #' 
 #' # convert sql file "sql.motus" to a tbl called "tbl.alltags"
 #' library(dplyr)
-#' tbl.alltags <- tbl(sql.motus, "alltags") 
+#' tbl.alltags <- tbl(sql.motus, "alltagsGPS") 
 #' 
 #' # convert the tbl "tbl.alltags" to a data.frame called "df.alltags"
 #' df.alltags <- tbl.alltags %>%