+++ R/tagSumSite.R
@@ -1,21 +1,24 @@
 #' Summarize detections of all tags by site
 #'
-#' Creates a summary for each tag of it's first and last detection time at each site,
-#' length of time between first and last detection of each site, and total number of detections at each site.
+#' Creates a summary for each tag of it's first and last detection time at each
+#' site, length of time between first and last detection of each site, and total
+#' number of detections at each site.
 #'
-#' @param data a selected table from .motus data, eg. "alltags", or a data.frame
-#'   of detection data including at a minimum variables for motusTagID, fullID,
-#'   recvDeployName, ts
+#' @param data a selected table from .motus data, eg. "alltagsGPS", or a
+#'   data.frame of detection data including at a minimum variables for
+#'   motusTagID, fullID, recvDeployName, ts, recvDeployLat, recvDeployLon,
+#'   gpsLat, gpsLon
 #' @param units units to display time difference, defaults to "hours", options
 #'   include "secs", "mins", "hours", "days", "weeks"
 #' @export
 #' @author Zoe Crysler \email{zcrysler@@gmail.com}
 #'
 #' @examples
-#' # You can use either a selected tbl from .motus eg. "alltags", or a
+#' # You can use either a selected tbl from .motus eg. "alltagsGPS", or a
 #' # data.frame, instructions to convert a .motus file to all formats are below.
 #' 
 #' # download and access data from project 176 in sql format
+#' # usename and password are both "motus.sample"
 #' \dontrun{sql.motus <- tagme(176, new = TRUE, update = TRUE)}
 #' 
 #' # use example sql file included in `motus`
@@ -24,7 +27,7 @@
 #' 
 #' # convert sql file "sql.motus" to a tbl called "tbl.alltags"
 #' library(dplyr)
-#' tbl.alltags <- tbl(sql.motus, "alltags") 
+#' tbl.alltags <- tbl(sql.motus, "alltagsGPS") 
 #' 
 #' # convert the tbl "tbl.alltags" to a data.frame called "df.alltags"
 #' df.alltags <- tbl.alltags  %>%