+++ R/tagme.R
@@ -86,8 +86,11 @@ tagme = function(projRecv, update = TRUE, new = FALSE, dir = getwd(),
              "If you *really* want to create a new database, specify 'new=TRUE'\n",
              "But maybe you just need to specify 'dir=' to tell me where to find it?"
              )
-    if (new && have)
-        warning("Database ", dbname, " already exists, so I'm ignoring the 'new=TRUE' option")
+    if (new && have) {
+        warning("Database ", dbname, " already exists, so I'm ignoring the ",
+                "'new=TRUE' option", immediate. = TRUE)
+        new <- FALSE
+    }
     if (new && missing(update))
         update = FALSE
     if (! have && is.character(projRecv)) {
@@ -109,13 +112,18 @@ tagme = function(projRecv, update = TRUE, new = FALSE, dir = getwd(),
         # - Passes and proceeds as expected
         if(!new) {
             rv <- checkDataVersion(rv, dbname = dbname, rename = rename)
+            # For receivers, if starting fresh, get the device ID again
+            if(length(DBI::dbListTables(rv$con)) == 0 && 
+               is.null(deviceID) && !is_proj(projRecv)) {
+                deviceID <- srvDeviceIDForReceiver(projRecv)[[2]]
+            }
         } else {
             # Prompt for authorization to update dataVersion prior to filling tables
             motus_vars$authToken 
         }
 
         # Ensure correct DBtables, but only if update = TRUE
-        ensureDBTables(rv, projRecv, deviceID)
+        ensureDBTables(rv, projRecv, deviceID, quiet = new)
 
         # Update databse
         rv <- motusUpdateDB(projRecv, rv, countOnly, forceMeta)